ofz#26944
this actually a regression from

    commit d38a79f386b234c062cbfe8ed81daf02851ce2fc
    Author: Noel <noelgrandin@gmail.com>
   Date:   Fri Oct 23 15:12:22 2020 +0200
    make SvXMLImport capable of mixing fast- and slow- contexts adhoc

where I did not get the namespace handling right, but it only
became obvious with

    commit 3adf150cbbf9c30d4a2429eb2f678cea81cb2b82
    Date:   Mon Nov 2 12:26:26 2020 +0200
    drop the SvXMLExport::EndElement method..

Specifically, we have weird logic to treat some bad namespaces
as good (who knows why), but that logic only exists for the
slowparser path.
With the dropping of EndElement(), we ended up with calls
to SvXMLImport::startUnknownElement() calling
SvXMLImportContext::StartElement(), but without
a corresponding call to SvXMLImportContext::endFastElement.

To make this work right, I copied the namespace aliasing code
to FastParser.

Change-Id: I00ecbf046feeaac6f2a789f801175dba40836f84
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105441
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>