Avoid changing anchor on visibility change of cell
This is an addition to commit a35709f93f4afa22fee7f41fcbc488992a16e642
The following scenario had produced a wrong object size: The object is
anchored to cell. Some columns containing this cell were hidden and
then shown again.
ScDrawLayer::SetCellAnchoredFromPosition was called in this case and
had produced the wrong size. When the column of the object anchor is
shown, object becomes visible. This gives an 'object change' event,
sent to ScDrawView::Notify, which calls adjustAnchoredPosition. That
had a test pAnchor->getShapeRect() == pObj->GetSnapRect() that should
prevent calling SetCellAnchoredFromPosition. But exact equality fails
due to +-1 differencies because of Twips<->Hmm conversions.

Change-Id: I0bd3684b7a5eda62b578275c02a5ac839ce58e2c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105802
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>