tdf#138039 tdf#134298 sw: layout: fix overlap of fly and table
The layout is horribly borked, the fly anchored in the body-level
paragraph messed with the preceding table:

page id="1" top="284" width="11905" height="16837" bottom="17120"
    tab id="3" top="794"
        row id="4" top="17121"
            fly id="8" top="16725"
    txt id="7" top="1394"
        fly ptr="0x6ce5510" id="10" top="1302"

SwTabFrame::CalcFlyOffsets() detects an overlap with the large fly, and
since it has wrap NONE it resizes to below the large image.

Then the SwTabFrame doesn't fit on the page, so it is split, but the split
fails because nDistanceToUpperPrtBottom is -720 (negative); hence it is
joined again.

Meanwhile the fly was invalidated, so now CalcFlyOffsets() ignores it and
the table shrinks again.

Once the fly is positioned again, the process repeats from the start.

Fix this in SwTabFrame::CalcFlyOffsets() by ignoring flys with wrap NONE that
extend below the body of the document and are anchored in a frame in the
next-chain of the table frame: these must move to the next page with their
anchor frame.

For the bugdoc this gives the same layout as LO 5.2.

Reportedly this problem started to happen since commit
94ea3ba58867a4f786903d0c617654b0c5e204b6, but it's not obvious why.

Change-Id: Iafb8a6afcba634f11c5db73869313ded0fe13bbd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105809
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>