sw replacement image creation: fix this when an ODT's settings.xml is invalid
This builds on top of commit 01da9cfbfceed33d05e77e4a759ffa8b7b52d7b4
(Bin overly eager early return that stops replacement image creation,
2020-08-20), and handles a similar case, when
SwView::ReadUserDataSequence() is not called at all.

The result is the same: no shell is selected on the command dispatcher
stack, so .uno:UpdateAll is ignored and the replacement images are not
created.

Change-Id: If8b74424b0ad88e63129fd8f7efd09f85e300895
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106068
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins