tdf#138010 (I) Pass bounding rect when natively drawing frame
Pass the bounding rectangle instead of the content rectangle
in the call to 'DrawNativeControl' when natively drawing a frame
here.

Passing the content rect instead would result in the
actual frame/border getting drawn into the area that is
supposed to be used by the actual content if
the content rect were different from the bounding rect.

As far as I can see, this commit should not make any practical
difference by itself for now, since Windows doesn't draw
'ControlType::Frame' natively, macOS's
'AquaSalGraphics::getNativeControlRegion' always returns the
same value for the bounding and the content rect and
gtk3 ('GtkSalGraphics::getNativeControlRegion') and
qt5 ('Qt5Graphics_Controls::getNativeControlRegion')
currently only take the border/frame width
into account to calculate the actual content rect when
the 'DrawFrameFlags::NoDraw' flag is passed, but
then the actual drawing is skipped here anyway.

This will be changed in follow-up commits to be
able to take the actual frame width into account
when rendering the actual content.

Change-Id: I57c29794c0a06ed7d6fd9be211dc427791554048
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106154
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>