Replace matrix translate with object Move
TR*BaseGeometry is faulty for SdrMeasureObj, fixing these methods is a
larger effort and Move is simpler anyway. So I replace the translation
using matrices with the Move methods for objects. Backup and restore
of the geometry values in state HiddenAsZero is now done with
SdrObjGeoData. That is known to work from use in Undo methods.

Change-Id: Ic32e8dbc5ae5f1a2f80e428448abc4b4a530b76b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106189
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>