oox: VML export: Output coordsize even when top/left is 0,0
For a polyline which is written as a shape with a path, the path
coordinates are defined by the coordsize; if it's missing we
end up writing hmm's when the reader is probably expecting 1000th
of parent.

Note groups are still very broken.
(e.g. try ./sw/qa/extras/ooxmlexport/data/kde216114-1.odt)

This fixes the offsetting seen in part 3 of tdf#138374

Change-Id: I61ba36527ee7716db0bda491df7fd9c7e4652dd1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106330
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>