sw: fix SwXTextRange::createEnumeration() inside table cell
This would set CursorType::SelectionInTable but leave m_pOwnTable
and m_pOwnStartNode uninitialised, causing

sw/source/core/unocore/unoobj2.cxx:399: SwXParagraphEnumeration: table type but no start node or table

and then the enumeration would return the table it's in as the first
element, which is quite annoying.

Refactor the creation of SwXParagraphEnumeration to prevent this.

Change-Id: I4e9e3456bdf66b9822d19ad985a20b094e6bbba4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106532
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>