with SAL_RTL_ENABLED=1 calc tabbar arrows are rendered reversed
in...

commit 96ee487ff5908ae6ce42f768c3af8e8e2df569d4
Date:   Fri Jul 17 17:04:27 2020 +0100

    tdf#134085 DrawTransformedBitmap in wrong place for RTL

    The return of mirror wasn't used, so no effort at a rtl mirror
    was made anyway.

I confidently claimed that "we don't want to mirror polygons", and we
definitely don't want to mirror images which was the immediately obvious
problem there. It was also the case that the highlight polygons drawn
there were missing entirely and the changes improved matters and put
those polygons at the correct position.

But it does seem that we *do* want to mirror polygons. Since...

commit 7e88c0bd3adec254b00c553848c2758e295b16c6
Date:   Fri Sep 25 13:30:11 2020 +0200

    disentangle AA and B2D use in VCL drawing

this path has become more used and makes it obvious that the triangles
in the tabbar buttons are reversed since the second commit.

Change-Id: Id709b2abff04c609a3c644530bcdb60c83a82942
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106702
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>