Very early work-in-progress commit for the "DevTools" dockable toolbar
This commit mostly just adds the plumbing to officecfg to display the
DevTools toolbar. The code for it does not yet implement any specific
functionality. The name "DevTools" is likely not what this will
actually be called in the end. The DevTools toolbar just contains two
unrelated combo boxes.

(See
https://blog.documentfoundation.org/blog/2020/07/27/tender-for-implementing-support-for-a-dedicated-built-in-uno-object-inspection-tool-in-libreoffice-202007-02/
)

Change-Id: I23cb129b48fde64ae39de5db5811cdf61c8993e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107432
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>