[API CHANGE] Do not call getaddrinfo if we just want the hostname
Similiar to 7e55caeefbf3f7c9e73055f08c88a7bacb7a8d91 ,
don't call getaddrinfo in getLocalHostname which calls DNS.

This causes a lag when creating the lockfile on opening a document
if the network is flaky/disabled.
See tdf#97931 and tdf#47179 for some problems caused by this.

For the one case where it is expected to call DNS, add a separate function
to restore the old behavior.

The (semantic) [API CHANGE] is in osl_getLocalHostname,
it does no longer return a FQDN.

Change-Id: I43455715a474ff6770351d1ce007c28aeb08f32e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107554
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@allotropia.de>