pdf: add a public function to load PDF as a VectorGraphicData
This adds a public function that loads a PDF as a VectorGraphicData
which essentially spilts the ImportPDF to general import as a
VectorGraphicData and then creation of a new Graphic. This is
needed so we can just load the VectorGraphicData independent to
a Graphic itself, so we don't needlessly create a new Graphic
instance in some cases.

Change-Id: Ib5f570242da69a1537bfdf1054f8eb40ecead31b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107991
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>