Sort custom entity names on fast parser
When there are lots of entities and a lot of cases to replace, it becomes a ballast to search the whole array.
So in order to avoid it, now uses ordered search and stops when OUString order implies that it can't be further.
The entity list is sorted before the parse by quick sort.

Change-Id: I9c91338ad67ddea1c273e329542549a904a0e563
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107774
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>