tdf#129961 svx: finish UI for table shadow as direct format
Normally properties on an SdrObject is set using SetAttributes(), but
that would take the selection controller into account, so we would call
SvxTableController::SetAttributes(), which sets the item set on the
selected cells instead. So use SetAttrToMarked() instead, which works on
the shape's item set, even in the table case. Don't replace all existing
items because we only have shadow properties here and also a disabled
shadow is still a (set) SdrOnOffItem (with value=false), so no old
SdrOnOffItem will be forgotten in the shape's item set.

Also add an outer undo grouping, so once the user presses OK in the
table properties dialog, we only create a single user-visible undo
action, not two.

Change-Id: I77b55ba1f07b8d0eeac5070e0ec07d39573d1f9a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107781
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins