tdf#136578 sw BreakDashedLine: edit/delete PageAfter
When you click on the blue dashed line indicating a page break,
that delete should remove both that page's break-before,
and the previous paragraph's break-after.
Also, edit should detect if a break comes from a PageAfter.

Prior to this commit, multiple things were wrong:
-the previous paragraph wasn't checked, so PageAfter
breaks were not deleted.
-edit looked at the wrong para if break was PageAfter.
-the current paragraph's PageAfter was deleted - deleting
the NEXT page break, and not THIS one.

I hope only tables are the exception to the rule.
It doesn't seem to matter which table pNd is set with the
NONE'd break. In theory, I could edit the real GetBreakItem
functions to add in that code, but that seems very
dangerous at this low of a level.

Prior commit 2105fa83b7615bbd30b886822117cb02757e1539
flattened this function and did other minor cleanup.

Change-Id: I9944c25853ec206f9febb6ea9f4e34c13c770ebb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108072
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>