Same SvStream reading fix as in commit [1] but for TypeSerializer
[1] 75dfc5143519091dec87154010b94e54100cb834

Explanation from the commit:
"to avoid chasing weird problems where we read past the end
of file, which leads to random data in the variable we read into.

I expect a couple of possible regressions from this change

(1) memory leaks caused by non-exception-safe memory handling.
Of which there should not be much because we're pretty good
about using smart pointer classes these days.

(2) Broken files which used to load, will no longer do so.
These will have to be debugged by putting a breakpoint
on the SvStreamEOFException constructor, and examining
the backtrace to see where we should be catching and ignoring
the exception to make the code continue to handle such broken
files."

Change-Id: I6f9ba7599c208b4340e86014e326e9a0693cd528
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108257
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>