tdf#104278: text in Type frame in DocInformation is not aligned left
As we can see from sw/uiconfig/swriter/ui/flddocinfopage.ui, we have a GtkTreeView in GtkFrame "typeframe", without the following property:
<property name="show_expanders">False</property>

A GtkTreeView seems have a default indentation if show_expanders is True (which is the default), see:
https://developer.gnome.org/gtk3/stable/GtkTreeView.html#gtk-tree-view-set-show-expanders

A "expander" would look like this: https://python-gtk-3-tutorial.readthedocs.io/en/latest/expander.html

I do not see a reason why we should use an expander here in this dialog. As a result, setting show_expanders to False will solve the problem.

Change-Id: I29cbb13f8d57aeb502079f278e4b620d97d8f170
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108261
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>