Return focus to edit window
commit 797915bf0bc0ed82c3cf58b7bd03c42d3335ca18 prevents focus from
returning to the edit window when Navigate by previous or next buttons
in the Navigator tool box are pressed. This may be nice for activation
of the button by keyboard as it keeps focus on the button, but it
doesn't allow for cursor indication of the navigated position in the
edit window.

Change-Id: I0067ee7d80aa94a510c83a52c22f788ba4a03dc6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108276
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>