vcl: remove OutputDevice::ScaleBitmap
Some time ago, in an attempt to remove a check on OutDevType I
refactored some code into the function ScaleBitmap(). When I look at
this more closely, I realise that I missed the point of the function.
The code I refactored actually checked if subsampled scaling could be
done. Thus, the function is inaccurate - it only scales *down* a bitmap,
and in fact scaling is already done in Bitmap.

Instead, I have now moved the function back into the code and replaced
it with a virtual function that checks if the OutputDevice derived class
can actually do bitmap subsampling (Printer is unable to do
subsampling).

Change-Id: I2e6bdbba0bf2153e0421efe6e05c05dd593a4e1e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108323
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>