tdf#138544 sw LoadUserSettings: default EmptyDbFieldHidesPara
This is the fourth patch in the series.
When you start a new document, Writer compat setting
"Hide paragraphs of database fields with an empty value" is on.
Now, when saving and reloading it with LoadUserSettings disabled,
it is still the program default of "on", instead of "off".

This one was a bit different because it didn't initialize
in the standard way.
    // for some properties we don't want to use the application
    // default if they're missing. So we watch for them in the loop
    // below, and set them if not found
So I'm changing it to match the standard way,
and adding the now familiar clause to cover the case
when the LoadUserSettings is disabled.

Change-Id: I979b01340000056567a4baa9ace27f9f9d15e9b4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108548
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>